webserver/requirements.txt
pytube>=12.1.0
openai-whisper>=20230314
streamlit>=1.25.0
supabase>=1.0.7
fastapi>=0.103.1
uvicorn>=0.23.2
pytesseract>=0.3.10
opencv-python>=4.8.1
yt-dlp>=2024.10.1
openai>=0.30.0
python-dotenv>=1.0.0
